==> retailrocket/01_Data_Source.ipynb <==
import pandas as pd
import os
import zipfile as zf
import getpass

### Install requirements

# Install kaggle
# https://www.kaggle.com/docs/api
!pip install kaggle
!mkdir -p ~/.kaggle

# Configure kaggle username and key
if not os.path.exists(os.path.expanduser('~/.kaggle/kaggle.json')):
    username = getpass.getpass("Username: ")
    key = getpass.getpass("Key: ")
    # Write credentials into file
    file = open(os.path.expanduser('~/.kaggle/kaggle.json'), "w")
    file.write("{\"username\":\"%s\",\"key\":\"%s\"}" %(username, key))
    file.close()
    !chmod 600 ~/.kaggle/kaggle.json

### Download dataset

# Download dataset
!kaggle datasets download --force -d retailrocket/ecommerce-dataset

# Extract dataset into data folder
files = zf.ZipFile("ecommerce-dataset.zip", "r")
files.extractall("data")
files.close()

# Remove zip file
os.remove("ecommerce-dataset.zip")

### Load dataset

events_df = pd.read_csv("data/events.csv")
category_tree_df = pd.read_csv("data/category_tree.csv")
item_properties_1_df = pd.read_csv("data/item_properties_part1.csv")
item_properties_2_df = pd.read_csv("data/item_properties_part2.csv")

events_df.head()

# Type of interactions / events
events_df.event.unique()

print(len(item_properties_1_df))
item_properties_1_df.head()

print(len(item_properties_2_df))
item_properties_2_df.head()

item_properties_df = pd.concat([item_properties_1_df, item_properties_2_df])
print(len(item_properties_df))
item_properties_df.head()

### Store Datasets

events_df.to_pickle("data/01_events_df.pkl")

category_tree_df.to_pickle("data/01_category_tree_df.pkl")

item_properties_df.to_pickle("data/01_item_properties_df.pkl")
