==> retailrocket/03_Data_Exploration.ipynb <==
import pandas as pd
import numpy as np

### Load datasets

events_prepared_df = pd.read_pickle("data/02_events_prepared_df.pkl")

train_sessions = pd.read_pickle("data/02_train_sessions.pkl")

test_sessions = pd.read_pickle("data/02_test_sessions.pkl")

### Data Exploration

* The **timestamp** portion is in Unix Epoch format e.g. 1433221332117 will be converted to Tuesday, 2 June 2015 5:02:12.117 AM GMT
* **visitorid** is the unique user currently browsing the website
* **event** is what the user is currently doing in that current timestamp
* **transactionid** will only have value if the user made a purchase as shown below

train_sessions.head()

# Type of interactions / events
print("Train interactions: ", train_sessions.event.unique())
print("Test interactions: ", test_sessions.event.unique())

print("Interactions without transactionid: ", train_sessions[train_sessions.transactionid.notnull()].event.unique())
print("Interactions with transactionid: ", train_sessions[train_sessions.transactionid.isnull()].event.unique())

# TODO
