==> retailrocket/05_Merlin_models.ipynb <==
import sys
import os
import glob
import gc

### Install requirements

# Install conda if not available
# curl https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -o /tmp/miniconda-installer.sh
# bash /tmp/miniconda-installer.sh

# Setup conda environment
# conda create -n rapids-22.02 -c rapidsai -c nvidia -c conda-forge python=3.8 cudatoolkit=11.2 cudf=22.02 dask-cudf=22.02
# conda activate rapids-22.02
# pip install nvtabular==0.11.0 tensorflow-gpu==2.8.0 merlin-models==0.2.0 transformers4rec==0.1.4 scipy==1.8.0 pynvml==11.4.1 ipykernel
# python -m ipykernel install --user --name=rapids-22.02

import pandas as pd
import numpy as np
import tensorflow as tf

import cudf
import dask_cudf

import nvtabular as nvt
from nvtabular.ops import *

from merlin.schema.tags import Tags
from merlin.schema import Schema

import merlin.models.tf as mm
import merlin.models.tf.dataset as tf_dataloader

from merlin.io.dataset import Dataset
from merlin.schema.io.tensorflow_metadata import TensorflowMetadata
from merlin.models.tf.blocks.core.aggregation import CosineSimilarity

from utils.fit_transform import workflow_fit_transform
from utils.save_visualize import save_results, print_results

output_path = './data/processed'
train_path = './data/train/*.parquet'
test_path = './data/test/*.parquet'

# Create data folders
!mkdir -p ./data/train ./data/test

# Define variables
TIME_KEY = "timestamp"
USER_KEY = "visitorid"
ITEM_KEY = "itemid"
SESSION_KEY = "sessionid"
TRANSACTION_KEY = "transactionid"
TARGET_KEY = "target"

### Load datasets

# Convert to parquet
train_sessions = pd.read_pickle("data/02_train_sessions.pkl")

# Introduce target variable
train_sessions[TARGET_KEY] = train_sessions[TRANSACTION_KEY].fillna(0)
train_sessions.loc[train_sessions[TARGET_KEY] != 0, TARGET_KEY] = 1
train_sessions[TARGET_KEY].value_counts()

train_sessions.to_parquet("data/train/02_train_sessions.parquet")

# Convert to parquet
test_sessions = pd.read_pickle("data/02_test_sessions.pkl")

# Introduce target variable
test_sessions[TARGET_KEY] = test_sessions[TRANSACTION_KEY].fillna(0)
test_sessions.loc[test_sessions[TARGET_KEY] != 0, TARGET_KEY] = 1
test_sessions[TARGET_KEY].value_counts()

test_sessions.to_parquet("data/test/02_test_sessions.parquet")

df = dask_cudf.read_parquet(train_path)
df.head()

df.columns

df[[USER_KEY, ITEM_KEY, TRANSACTION_KEY, TARGET_KEY]].head()

# Check the total number or rows
len(df)

# Print item stats
print('Minimum item id: ' + str(df[ITEM_KEY].min().compute()))
print('Maximum item id: ' + str(df[ITEM_KEY].max().compute()))
print('Unique item id: ' + str(df[ITEM_KEY].unique().shape[0].compute()))

# Print user stats
print('Minimum user id: ' + str(df[USER_KEY].min().compute()))
print('Maximum user id: ' + str(df[USER_KEY].max().compute()))
print('Unique user id: ' + str(df[USER_KEY].unique().shape[0].compute()))

# Check the distribution of the target class. We can see that the dataset is imbalanced.
df[TARGET_KEY].value_counts().compute()

### Merlin Models

[Merlin Models](https://github.com/NVIDIA-Merlin/models) is a library to make it easy for users in industry or academia to train and deploy recommender models with best practices baked into the library. This will let users in industry easily train standard models against their own dataset, getting high performance GPU accelerated models into production. This will also let researchers to build custom models by incorporating standard components of deep learning recommender models, and then benchmark their new models on example offline datasets.

%%time

## Define NVT the pipeline
# We add tags for user_id, item_id and target. NVTabular will provide an output file
# We categorify the user_id and item_id to be continuous integers 0, ..., |C|
user_id = [USER_KEY] >> AddMetadata(tags=[Tags.USER_ID]) >> Categorify(freq_threshold=5)
item_id = [ITEM_KEY] >> AddMetadata(tags=[Tags.ITEM_ID]) >> Categorify(freq_threshold=5)
targets = [TARGET_KEY] >> AddMetadata(
    tags=[str(Tags.BINARY_CLASSIFICATION), TARGET_KEY]
)

# Add more features
add_feat = ["event"] >> nvt.ops.Categorify()

# Add target encoding
te_feat = (
    [USER_KEY, ITEM_KEY] + add_feat >>
    TargetEncoding(
        [TARGET_KEY],
        kfold=1,
        p_smooth=20
    ) >>
    Normalize()
)

outputs = user_id + item_id + targets + add_feat + te_feat

etl_description = 'etl'

%%time

# Run NVTabular workflow
workflow_fit_transform(outputs, train_path, test_path, output_path)

# Load schema
schema = TensorflowMetadata.from_proto_text_file(output_path + '/train/').to_merlin_schema()

target_column = schema.select_by_tag(Tags.TARGET).column_names[0]

# Define model
model = mm.DCNModel(
    schema,
    depth=2,
    deep_block=mm.MLPBlock([64, 32]),
    prediction_tasks=mm.BinaryClassificationTask(target_column, 
                                                 metrics=[tf.keras.metrics.AUC(), tf.keras.metrics.Accuracy(),
                                                         tf.keras.metrics.Precision(), tf.keras.metrics.Recall()])
)
model_description = "DCN"

model.compile(optimizer="adam", run_eagerly=False)
model_description = "DCN_adam"

batch_size = 16 * 1024

train_dl = tf_dataloader.BatchedDataset(
    Dataset(output_path + '/train/*.parquet', part_size="500MB"),
    batch_size = batch_size,
    label_names = [target_column],
    shuffle= True,
    schema = schema,
)

test_dl = tf_dataloader.BatchedDataset(
    Dataset(output_path + '/test/*.parquet', part_size="500MB"),
    batch_size = batch_size,
    label_names = [target_column],
    shuffle = False,
    schema = schema,
)

%%time
model.fit(train_dl, validation_data=test_dl)

# model_des = model_description + '_' + etl_description
for key, value in  model.history.history.items():
    print('%s:%s' % (key, value[0]))

### Cleanup

# Let's delete the dataframe to free GPU-memory
del df
gc.collect()

# We now shutdown the notebook to free GPU-memory
import IPython
app = IPython.Application.instance()
app.kernel.do_shutdown(True)
